Clean set lang types
