+++ src/utils/constants/index.ts
@@ -4,7 +4,6 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-export * from './browserConstants';
 export * from './fetchStatus';
 export * from './fieldConstants';
 export * from './uiConstants';
