+++ src/utils/index.ts
@@ -8,6 +8,7 @@ export * from './algos';
 export * from './constants';
 export * from './conversionUtils';
 export * from './functions';
+export * from './langs';
 export * from './mapper';
 export * from './styles';
 export * from './types';
