+++ src/utils/langs.ts
@@ -1,11 +1,12 @@
-/**
- * Copyright (c) 2024, RTE (http://www.rte-france.com)
+/*
+ * Copyright © 2025, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-export const DARK_THEME = 'Dark';
-export const LIGHT_THEME = 'Light';
+
 export const LANG_SYSTEM = 'sys';
 export const LANG_ENGLISH = 'en';
 export const LANG_FRENCH = 'fr';
+export type GsLangUser = typeof LANG_ENGLISH | typeof LANG_FRENCH;
+export type GsLang = GsLangUser | typeof LANG_SYSTEM;
